I seem to have finally really fixed responders dealing with console commands by modifying G_Responder. Added a beta version of joe's steam spawner and repaired bugs in it to help him out. Finished polyobject sound sequence support.

+++ eternity/source/s_sndseq.h
@@ -97,7 +97,7 @@ typedef struct SndSeq_s
 
    // 10/17/06: data needed for savegames
    int originType;               // type of origin (sector, polyobj, other)
-   ptrdiff_t originIdx;          // sector or polyobj number, (or -1)
+   int originIdx;                // sector or polyobj number, (or -1)
 } SndSeq_t;
 
 // Sound sequence pointers, needed for savegame support
